Add SimpleNumberField.kt Edit PasswordField.kt - Make keyboard for password

+++ app/src/main/java/at/ac/fhcampuswien/budget_fox/models/Income.kt
@@ -5,7 +5,7 @@ import java.util.UUID
 
 data class Income(
     val uuid: UUID = UUID.randomUUID(),
-    val amount : Float,
+    val amount : Double,
     val description: String,
     val period: Period? = null // made period nullable to enable non periodical incomes
 )
